simplify(S04): move bd/jq work-query shell codegen to workquery.go [a]
Phase 1 (mechanical, verbatim, zero API change): move the ~700-line
bd/jq shell-codegen block plus the Effective*Query accessor triplets out
of the most-churned file in the repo (config.go) into a new same-package
file internal/config/workquery.go.

Sites changed:
- internal/config/config.go: removed lines 3489-4196 (bdReadyPoolDemandShell
  through effectiveOnBoot) and dropped the now-unused beadmeta + shellquote
  imports; 5183 -> 4473 lines.
- internal/config/workquery.go: new file holding the moved block verbatim
  with its required imports (fmt, strconv, strings, time, beadmeta,
  shellquote).

Same-package move: unexported helpers and exported Effective*Query names
are unchanged, so all consumers (cmd/gc, internal/materialize,
internal/graphroute) compile untouched.

Gates: build=pass vet=pass tests=pass
